first

+++ src/Taska.java
@@ -1,17 +1,17 @@
 import java.util.Scanner;
+
 public class Taska {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
 
         double c = Math.sqrt(a * a + b * b);
 
         System.out.println(c);
     }
 }
-
